+++ src/index.html
@@ -18,7 +18,7 @@
 
     </div>
     <div class="clear">
-        <h3>Clear all completed</h3>
+        <h3 class="clear-data">Clear all completed</h3>
     </div>
     </div>
 </body>
